+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/entities/nonliving/PurpleSpikeEntity.java
@@ -189,7 +189,7 @@ public class PurpleSpikeEntity extends Entity {
 
                             livingEntity.addEffect(new MobEffectInstance(
                                     MobEffects.POISON,
-                                    200,
+                                    50,
                                     1,
                                     true,
                                     true,