+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+### **(V.7.15.4 Changes) (1.21.1 Minecraft)**
+
+#### Entities:
+Tiny optimization for Sentry Watcher/Cosmic Crystal.
+
+Reduced duration of Purple Spike's poison effect.
+
+#### Items:
+Fixed how damage is calculated for thrown Stinger Spear and ensure it always poisons non-undead mobs.
+
+Fixed Carpenter Bee Boots to give mining boost from boot's Efficiency enchantments, not whatever is in hand.
+
+Carpenter Bee Boots tooltip now clarifies mining enchantments applies to when mining by feet.
+
+Added Trials Spawners and Vaults as blocks that Knowing Essence can highlight.
+
+Fixed thrown Dirt Pellet not being in `minecraft:redirectable_projectile` entity type tag.
+
+#### Structures:
+Slightly reduced enemy mob speed boosts from Blue and Red Essence Arenas.
+
+
 ### **(V.7.15.3 Changes) (1.21.1 Minecraft)**
 
 #### Fluids: